requirements.txt
fastapi
uvicorn

sqlalchemy
mysqlclient
PyMySQL
PyMySQL[rsa]

python-dotenv
python-jose

pytest
requests
pytest-cov

cryptography

cpf-generator
